\name{equateMultiple-internal}
\alias{irtp1}
\alias{obj}
\alias{Sgamma}
\alias{pABgammaR2C}
\alias{pABgammaR}
\alias{hessR}
\alias{gamma2ab}
\alias{multiec_irf}
\alias{ipfR}
\alias{der_A_discrmn_bis}
\alias{der_asAbsB_ab}
\alias{multiec_moments}
\alias{matD}
\alias{Dmat}

\title{Internal Functions of Package equateIRT}
\description{
  Internal Functions of Package equateIRT.
}
\details{
  These are not to be called by the user.
}
\keyword{ internal }
